Refactor: Use Int instead of Byte for Card type in Range
* Update Card type to Int
* Replace all usages of Byte with Int in card-related operations
* Adapt functions `charToRank`, `charToSuit`, `rankToChar`, `suitToChar`, `getWeightPair`, `getWeightSuited`, and `getWeightOffsuit` to use Int instead of Byte.
* Modify the card pair and index calculation to utilize Int instead of Byte
* Update related tests to work with Int instead of Byte.

+++ presentation/screens/src/commonTest/kotlin/com/bz/movies/presentation/screens/postflop/RangeTest.kt
@@ -14,9 +14,9 @@ class RangeTest {
         for (i in 0..51) {
             for (j in (i + 1)..51) {
                 println("k:$k , i:$i , j:$j")
-                assertEquals(actual = cardPairToIndex(i.toByte(), j.toByte()), expected = k)
-                assertEquals(actual = cardPairToIndex(j.toByte(), i.toByte()), expected = k)
-                assertEquals(actual = indexToCardPair(k), expected = i.toByte() to j.toByte())
+                assertEquals(actual = cardPairToIndex(i, j), expected = k)
+                assertEquals(actual = cardPairToIndex(j, i), expected = k)
+                assertEquals(actual = indexToCardPair(k), expected = i to j)
                 k += 1
             }
         }
